# src/IO.Swagger/Controllers/TransacaoApi.cs
/*
 * ByteBank Api
 *
 * Administra a movimentação bancária do correntista e controla o acesso através de CPF e senha.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: suporte@bytebankapi.com
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;
using IO.Swagger.Security;
using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class TransacaoApiController : ControllerBase
    { 
        /// <summary>
        /// Altera o Valor da Transacao
        /// </summary>
        /// <remarks>Altera o valor da transação existente na conta do cliente</remarks>
        /// <param name="codigo"></param>
        /// <param name="valor"></param>
        /// <param name="authorization"></param>
        /// <response code="204">Alteração do valor efetuada com sucesso</response>
        /// <response code="400">Requisição inválida</response>
        /// <response code="401">Requisição não autorizada</response>
        /// <response code="500">Erro no servidor</response>
        [HttpPut]
        [Route("/bytebank-api/v1/trasacao/{codigo}")]
        [Authorize(AuthenticationSchemes = BasicAuthenticationHandler.SchemeName)]
        [ValidateModelState]
        [SwaggerOperation("AlteraValorTransacao")]
        public virtual IActionResult AlteraValorTransacao([FromRoute][Required]long? codigo, [FromQuery][Required()]double? valor, [FromHeader][Required()]string authorization)
        { 
            //TODO: Uncomment the next line to return response 204 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(204);

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);

            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500);


            throw new NotImplementedException();
        }

        /// <summary>
        /// Cadastra Transação na Conta
        /// </summary>
        /// <remarks>Cadastra uma transação de débito ou depósito na conta</remarks>
        /// <param name="authorization"></param>
        /// <param name="transacao"></param>
        /// <param name="tipo"></param>
        /// <response code="201">Cadastro de transação efetuado com sucesso.</response>
        /// <response code="400">Requisição inválida</response>
        /// <response code="401">Requisição não autorizada</response>
        /// <response code="500">Erro no servidor</response>
        [HttpPost]
        [Route("/bytebank-api/v1/transacao")]
        [Authorize(AuthenticationSchemes = BasicAuthenticationHandler.SchemeName)]
        [ValidateModelState]
        [SwaggerOperation("CadastraTransacao")]
        [SwaggerResponse(statusCode: 201, type: typeof(Transacao), description: "Cadastro de transação efetuado com sucesso.")]
        public virtual IActionResult CadastraTransacao([FromHeader][Required()]string authorization, [FromBody]Transacao transacao, [FromQuery][Required()]string tipo)
        { 
            //TODO: Uncomment the next line to return response 201 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(201, default(Transacao));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);

            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500);

            string exampleJson = null;
            exampleJson = "{\n  \"tipo\" : \"debito\",\n  \"data\" : \"2000-01-23T04:56:07.000+00:00\",\n  \"valor\" : 6.027456183070403,\n  \"id\" : 0\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<Transacao>(exampleJson)
            : default(Transacao);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Exclui a Transação
        /// </summary>
        /// <remarks>Exclusão da transação existente na conta do cliente</remarks>
        /// <param name="codigo"></param>
        /// <param name="authorization"></param>
        /// <response code="204">Transação excluída com sucesso</response>
        /// <response code="400">Requisição inválida</response>
        /// <response code="401">Requisição não autorizada</response>
        /// <response code="500">Erro no servidor</response>
        [HttpDelete]
        [Route("/bytebank-api/v1/trasacao/{codigo}")]
        [Authorize(AuthenticationSchemes = BasicAuthenticationHandler.SchemeName)]
        [ValidateModelState]
        [SwaggerOperation("ExcluiTransacao")]
        public virtual IActionResult ExcluiTransacao([FromRoute][Required]long? codigo, [FromHeader][Required()]string authorization)
        { 
            //TODO: Uncomment the next line to return response 204 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(204);

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);

            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500);


            throw new NotImplementedException();
        }
    }
}